Added many setThreadSafeRefUnref. Seems that it no longer crash at exit on non-animated models. But it only seems, maybe error case just appears less often.
Also there is a strange thing -- some shaders are compiled on program exit. They compile logs are interleaved with 
"Waiting for OperationThread to cancel ..". This this appear when model is moved so that not all meshes are visible 
in viewer. Looks like some shaders are destroyed before they are used again (for drawing invisible meshes???).


git-svn-id: https://svn.code.sf.net/p/osgcal/code/branches/0.3.0-dl@121 4711c375-e22e-0410-9441-d0ab9e8e3247